created music model and music controller
