+++ Model/UserModel.php
@@ -4,18 +4,6 @@ require "/Applications/XAMPP/xamppfiles/htdocs/inc/bootstrap.php";
 
 class UserModel extends Database
 {
-/**
- * The getUsers function retrieves a specified number of users from the database.
- * 
- * @param limit The limit parameter is used to specify the maximum number of rows to be returned from
- * the database query.
- * 
- * @return The getUsers function is returning the result of a database query. 
- */
-    public function getUsers($limit)
-    {
-        return $this->select("SELECT * FROM users LIMIT ?", ["i", $limit]);
-    }
 
 /**
  * The function retrieves the username and password of a user from the database based on their
